dfaulo_tool/requirements.txt
opencv-python~=4.5.2.52
numpy~=1.21.6
torch~=1.13.1+cu116
tensorflow~=2.8.0
Pillow~=9.0.1
matplotlib~=3.4.1
pyod~=0.9.9
torchvision~=0.14.1+cu116
sklearn~=0.0
scikit-learn~=0.24.2
tqdm~=4.61.1
pandas~=1.2.5
torchtext~=0.14.1
xlwt~=1.3.0
cleanlab~=2.0.0
keras~=2.8.0
skorch~=0.11.0
scipy~=1.10.1
xlrd~=1.2.0
pynput~=1.7.6